+++ NEWS.md
@@ -1,8 +1,12 @@
 # hpa.spatial (development version)
 
-* allow `get_correspondence_tbl()` to take a `from_geo` and `to_geo` rather than
+* allow `map_data_with_correspondence()` to take `from_geo` and `to_geo` rather
+  than only areas and years. (Also requires adding `export_fname` arg, 
+  consistent with changes to `get_correspondence_tbl()`.)
+
+* allow `get_correspondence_tbl()` to take `from_geo` and `to_geo` rather than
   only areas and years to get the polygons via `get_polygon()`. Also allow user
-  to specify an `export_fname` (filename) to save the correspondence table for 
+  to specify an `export_fname` (file name) to save the correspondence table for 
   these new tables as they can't be named with `make_fname()`
 
 # hpa.spatial 0.2.6